refactor: moved ExternalConnector related stuff to common.edc
