+++ src/main/java/org/eclipse/tractusx/puris/backend/common/edc/domain/model/ExternalConnector.java
@@ -18,15 +18,32 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.eclipse.tractusx.puris.backend.model.repo;
+package org.eclipse.tractusx.puris.backend.common.edc.domain.model;
 
-import org.eclipse.tractusx.puris.backend.model.ExternalConnector;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import jakarta.persistence.Entity;
+import org.eclipse.tractusx.puris.backend.model.JpaBaseEntity;
 
 /**
- * JPA Repository for {@link ExternalConnector} entities.
+ * External Connector object, will be used for dropdown selection in frontend.
  */
-@Repository
-public interface ExternalConnectorRepository extends JpaRepository<ExternalConnector, Long> {
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class ExternalConnector extends JpaBaseEntity {
+
+    /**
+     * Display name of the connector.
+     */
+    private String name;
+
+    /**
+     * URL of the connector.
+     */
+    private String url;
+
 }
